Changes made to proto and ObjectRepository with respect to datatypes.

+++ src/main/java/org/anonymous/module/ObjectRepository.java
@@ -474,11 +474,11 @@ public class ObjectRepository implements AutoCloseable {
                 try {
                     Metadata metadata = Metadata.newBuilder().setSecurityName(rs.getString("name"))
                             .setSecurityType(rs.getInt("typeId"))
-                            .setUpdateCount(rs.getInt("updateCount"))
+                            .setUpdateCount(rs.getLong("updateCount"))
                             .setDateCreated(rs.getInt("dateCreated"))
                             .setTimeUpdate(rs.getString("timeUpdated"))
                             .setDbIdUpdated(rs.getInt("dbIdUpdated"))
-                            .setLastTxnId(rs.getInt("lastTransaction"))
+                            .setLastTxnId(rs.getLong("lastTransaction"))
                             .setVersionInfo(rs.getInt("versionInfo")).build();
                     CmdGetManyByNameExtResponse.ResponseMessage.MsgOnSuccess msgOnSuccess = CmdGetManyByNameExtResponse.ResponseMessage.MsgOnSuccess.newBuilder()
                             .setMetadata(metadata)
@@ -516,11 +516,11 @@ public class ObjectRepository implements AutoCloseable {
                 try {
                     Metadata metadata = Metadata.newBuilder().setSecurityName(rs.getString("name"))
                             .setSecurityType(rs.getInt("typeId"))
-                            .setUpdateCount(rs.getInt("updateCount"))
+                            .setUpdateCount(rs.getLong("updateCount"))
                             .setDateCreated(rs.getInt("dateCreated"))
                             .setTimeUpdate(rs.getString("timeUpdated"))
                             .setDbIdUpdated(rs.getInt("dbIdUpdated"))
-                            .setLastTxnId(rs.getInt("lastTransaction"))
+                            .setLastTxnId(rs.getLong("lastTransaction"))
                             .setVersionInfo(rs.getInt("versionInfo")).build();
                     CmdGetManyByNameExtResponseStream.MsgOnSuccess msgOnSuccess = CmdGetManyByNameExtResponseStream.MsgOnSuccess.newBuilder()
                             .setMetadata(metadata)
